Match plugin folder structure
