+++ index.js
@@ -1,0 +1,1 @@
+export { default as grid } from './src/d3-grid';
